+++ bmff/bufReader.go
