+++ bmff/ftyp.go
