+++ bmff/itemInfo.go
