+++ exif/exif.go
@@ -5,16 +5,16 @@ import (
 	"bufio"
 	"io"
 
+	"github.com/evanoberholster/imagemeta"
 	"github.com/evanoberholster/imagemeta/exif/ifds"
 	"github.com/evanoberholster/imagemeta/imagetype"
-	"github.com/evanoberholster/imagemeta/metadata"
 )
 
 // Errors
 var (
 	// Alias to meta Errors
-	ErrInvalidHeader = metadata.ErrInvalidHeader
-	ErrNoExif        = metadata.ErrNoExif
+	ErrInvalidHeader = imagemeta.ErrInvalidHeader
+	ErrNoExif        = imagemeta.ErrNoExif
 )
 
 // ScanExif identifies the imageType based on magic bytes and
@@ -36,7 +36,7 @@ func ScanExif(r io.ReaderAt) (e *ExifData, err error) {
 
 	// Search Image for Metadata Header using
 	// Imagetype information
-	m, err := metadata.ScanBuf(br, it)
+	m, err := imagemeta.ScanBuf(br, it)
 	if err != nil {
 		if err != ErrNoExif {
 			return
@@ -73,7 +73,7 @@ func ParseExif(r io.ReaderAt) (e *ExifData, err error) {
 
 	// Search Image for Metadata Header using
 	// Imagetype information
-	m, err := metadata.ScanBuf(br, imagetype.ImageUnknown)
+	m, err := imagemeta.ScanBuf(br, imagetype.ImageUnknown)
 	if err != nil {
 		if err != ErrNoExif {
 			return
