+++ imagemeta.go
@@ -1,2 +1,3 @@
 // Package imagemeta provides functions for parsing and extracting Metadata from Images.
+// Different image types such as JPEG, Camera Raw, DNG, TIFF, HEIF, and AVIF.
 package imagemeta
