--- metadata/xmp.go
@@ -1,1 +1,0 @@
-package metadata
